+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix incorrect `undefined_name` errors when a class is nested in a nested
+  function and uses a name from the outer function (#750)
 - Fix incorrect `possibly_undefined_name` error on certain uses of the
   walrus operator (#749)
 - Fix narrowing on `isinstance` calls with arguments that are not